un-pub `GlobalPoolAlreadyInitialized` by moving to registry.rs
